plugins/sr-csv/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sr-csv"
version = "0.1.0"
description = "CSV source plugin for Narrowfield"
license = "MIT"
requires-python = ">=3.10"
dependencies = ["narrowfield-sdk>=0.1.0"]

[project.urls]
Homepage = "https://github.com/insomniak68/narrowfield-plugins"
